backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
scipy
coqui-tts
torch
librosa
